backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
openai>=1.12.0
httpx>=0.27.0
pydantic==2.5.0
pydantic[email]==2.5.0
gunicorn==21.2.0
slowapi>=0.1.9

